modifier image detection script and scene
now every time retry button is clicked AR session is restarted. this is done because when arfoundation detects one image and switches to another image  , it cant detect the first image again.

AR Foundation v5.1.1
There is no API to determine the visibility of an image. Generally, if the tracking state is Tracking, it will likely change to Limited when the image is not visible. However, there are other situations in which the tracking state can be in states other than Tracked